handle ssr issue
