+++ lib/browser/webgl.js
@@ -1,0 +1,13 @@
+export function isWebglSupported() {
+  try {
+    const canvas = document.createElement('canvas')
+
+    return Boolean(window.WebGLRenderingContext) && (
+      canvas.getContext('webgl') || canvas.getContext('experimental-webgl')
+    )
+  } catch (err) {
+    return false
+  }
+}
+
+export default isWebglSupported
